requirements.txt
Flask==2.1.1
authlib==1.0.0
python-dotenv==0.20.0